+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
+## [0.0.3] - 2019-03-04
+
+### Added
+
 - `updateTextElements` function to update and reflow text with `force` parameter
 - `getNode`, `getEdge`, `getEdgesBySource` and `getEdgesByTarget` functions in grapheditor